+++ src/fat32.h
@@ -207,11 +207,11 @@
    * @brief   Get File Info from Root Directory
    *
    * @param   FAT32_t * FAT32
-   * @param   uint32_t file number
+   * @param   uint8_t file number
    *
    * @return  DE_t *
    *  */
-  DE_t * FAT32_Get_File_Info (FAT32_t *, uint32_t);
+  DE_t * FAT32_Get_File_Info (FAT32_t *, uint8_t);
 
   /**
    * @brief   Read Next Cluster From FAT
